requirements.txt
asgiref==3.5.0
Django==3.2
djangorestframework==3.13.1
passlib==1.7.4
pytz==2022.1
sqlparse==0.4.2
